+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Expected: October 2025
 * New: Compare RPC as defined in RFC9144 and yang module ietf-nmda-compare
 * New: [NETCONF and RESTCONF Private candidate datastores](https://github.com/clicon/clixon/issues/615)
 * All internal session clients start with NETCONF hello
+* New `clixon-lib@2025-08-01.yang` revision
+   * Added `netconf-framing-type`
 * New `clixon-config@2025-08-01.yang` revision
    * Added option: `CLICON_XMLDB_CANDIDATE_INMEM` to run candidate datastore in-mem
    * Added option: `CLICON_NETCONF_MONITORING_GETSCHEMA_CDATA` to encode schema content with CDATA